more cleanup & work on event system & allocators

+++ include/atma/utf/utf8_char.hpp
@@ -95,7 +95,7 @@ namespace atma
 	}
 
 	// return how many bytes we need to advance, assuming we're at a leading byte
-	inline auto utf8_char_bytecount(char const* leading) -> int8
+	inline auto utf8_char_bytecount(char const* leading) -> int
 	{
 		ATMA_ASSERT(leading);
 		ATMA_ENSURE(utf8_byte_is_leading(*leading));
